refactor(cli): split main.go into per-resource command files
